backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pinecone-client>=2.2.4
sentence-transformers>=2.7.0
torch>=2.2.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
google-generativeai>=0.8.0
